MAINT: Delete error_computation.py and test_error_computation.py
